backend/pyproject.toml
[tool.poetry]
name = "transparencia-fiscal"
version = "1.0.0"
description = "Plataforma RAG Híbrido para Transparência Fiscal"
authors = ["André Ferreira Santana"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.35"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
pgvector = "^0.3.6"
langchain = "^0.3.7"
langchain-anthropic = "^0.3.0"
langchain-community = "^0.3.5"
anthropic = "^0.39.0"
sentence-transformers = "^3.3.0"
httpx = "^0.27.0"
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
python-dotenv = "^1.0.1"
tenacity = "^9.0.0"
numpy = "^1.26.0"
pandas = "^2.2.0"
structlog = "^24.4.0"
sqlparse = "^0.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
ruff = "^0.8.0"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
